token implemented

+++ src/pages/services/userService.js
@@ -1,10 +1,10 @@
 import axios from "axios";
 
 const BASE_URL = "https://vetrasales-backend-production.up.railway.app/api/user";
-
+const token = localStorage.getItem('token')
 export const getAllUser = async ()=>{
     try {
-        const res = await axios.get(`${BASE_URL}/getAllUser`,{headers: { "Content-Type": "application/json" }});
+        const res = await axios.get(`${BASE_URL}/getAllUser`,{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
         console.log("API response:", res.data); // 👈 DEBUG
         return res.data
     } catch (error) {
@@ -16,7 +16,7 @@ export const getAllUser = async ()=>{
 
 export const getAllCustomer = async ()=>{
     try {
-        const res = await axios.get(`${BASE_URL}/getAllCustomer`,{headers: { "Content-Type": "application/json" }});
+        const res = await axios.get(`${BASE_URL}/getAllCustomer`,{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
         return res.data
     } catch (error) {
       console.error("Error fetching products:", error);
@@ -27,7 +27,7 @@ export const getAllCustomer = async ()=>{
 
 export const getAllAdmin = async ()=>{
     try {
-        const res = await axios.get(`${BASE_URL}/getAllAdmin`,{headers: { "Content-Type": "application/json" }});
+        const res = await axios.get(`${BASE_URL}/getAllAdmin`,{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
         return res.data
     } catch (error) {
       console.error("Error fetching products:", error);
@@ -37,7 +37,7 @@ export const getAllAdmin = async ()=>{
 }
 export const searchUser = async (name) => {
     try {
-        const res = await axios.get(`${BASE_URL}/searchUser`,{param:{name}},{headers: { "Content-Type": "application/json" }});
+        const res = await axios.get(`${BASE_URL}/searchUser`,{param:{name}},{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
         console.log("API response:", res.data); // 👈 DEBUG
         return res.data;
     } catch (error) {
@@ -50,20 +50,20 @@ export const searchUser = async (name) => {
 export const deleteUser = (id) => {
   return axios.get(`${BASE_URL}/delete`, {
     params: { id }
-  },{headers: { "Content-Type": "application/json" }});
+  },{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
 };
 
 export const signUp = (user) =>
   axios.post(`${BASE_URL}/signUp`, user, {
-    headers: { "Content-Type": "application/json" },
+    headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` },
   });
 
-export const signIn = (login) => axios.post(`${BASE_URL}/signIn`,login,{headers: { "Content-Type": "application/json" }});
+export const signIn = (login) => axios.post(`${BASE_URL}/signIn`,login,{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
 export const userData = async (username) => {
     try {
         const res = await axios.get(`${BASE_URL}/userData`,{
         params: { username },   // ✅ REQUIRED
-      },{headers: { "Content-Type": "application/json" }});
+      },{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
         console.log("API response:", res.data); // 👈 DEBUG
         return res.data;
     } catch (error) {
@@ -76,7 +76,7 @@ export const contact = async (username) => {
     try {
         const res = await axios.get(`${BASE_URL}/getEmail`,{
         params: { username },   // ✅ REQUIRED
-      },{headers: { "Content-Type": "application/json" }});
+      },{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
         console.log("API response:", res.data); // 👈 DEBUG
         return res.data;
     } catch (error) {
@@ -84,11 +84,11 @@ export const contact = async (username) => {
       return []; // prevent undefined
   }
 }
-export const deleteAll = () => axios.get(`${BASE_URL}/deleteAll`,{headers: { "Content-Type": "application/json" }});
-export const updateUser = (user) => axios.post(`${BASE_URL}/updateUser`,user,{headers: { "Content-Type": "application/json" }});
+export const deleteAll = () => axios.get(`${BASE_URL}/deleteAll`,{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
+export const updateUser = (user) => axios.post(`${BASE_URL}/updateUser`,user,{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
 export const emails = async () => {
     try {
-        const res = await axios.get(`${BASE_URL}/emails`,{headers: { "Content-Type": "application/json" }});
+        const res = await axios.get(`${BASE_URL}/emails`,{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
         console.log("API response:", res.data); // 👈 DEBUG
         return res.data;
         } catch (error) {
@@ -98,7 +98,7 @@ export const emails = async () => {
 }
 export const phones = async () => {
     try {
-        const res = await axios.get(`${BASE_URL}/phones`,{headers: { "Content-Type": "application/json" }});
+        const res = await axios.get(`${BASE_URL}/phones`,{headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }});
         console.log("API response:", res.data); // 👈 DEBUG
         return res.data;
     } catch (error) {
@@ -110,5 +110,5 @@ export const phones = async () => {
 export const toggleStatus = (id) =>
   axios.get(`${BASE_URL}/userStatus`, {
     params: { id },
-    headers: { "Content-Type": "application/json" }
+    headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` }
   });
